demos/dynamic_discovery_with_agent_cards/pyproject.toml
[project]
name = "status-update-agent-with-dynamic-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.5",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
